feat: Implement usd converter
